Fallback to previous HAL version

+++ Src/application/solverMaze/robotInterface.c
@@ -106,7 +106,7 @@ void move_zhonx(char direction_to_go, positionRobot *positionZhonx, unsigned int
         bluetoothWaitReady();
         bluetoothPrintf("start cell ");
 #endif
-        mazeMoveStartCell(max_speed_translation, min_speed_translation);
+        mazeMoveStartCell((double)max_speed_translation, (double)min_speed_translation);
         numberOfCell--;
     }
     else // so endMidOfCase=TRUE and positionZhonx->midOfCase=FALSE
